datahub_v3_project/requirements.txt
amqp==5.1.1  
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2    
billiard==3.6.4.0  
backports.zoneinfo;python_version<"3.9"
celery==5.1.2
certifi==2022.9.24
cffi==1.15.1 
charset-normalizer==2.1.1
click==7.1.2    
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3   
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-filter==22.1 
drf-yasg==1.21.4
filelock==3.8.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.23.5
oscrypto==1.3.0
packaging==21.3
pandas==1.5.1
pip==23.0
prompt-toolkit==3.0.32
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==1.7.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
setuptools==65.3.0
six==1.16.0
snowflake-connector-python==2.8.2
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
wrapt==1.14.132
celery==5.1.2
